pyproject.toml
[project]
name = "phonikud-chatterbox-runpod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "phonikud-chatterbox",
    "pydub>=0.25.1",
    "runpod>=1.7.13",
    "soundfile>=0.13.1",
]

[tool.uv.sources]
phonikud-chatterbox = { git = "https://github.com/thewh1teagle/phonikud-chatterbox" }
